tokenswave/requirements.txt
asgiref==3.8.1
django
gunicorn==23.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
